requirements.txt
ultralytics
streamlit
opencv-python-headless
Pillow
google-generativeai
python-dotenv
numpy
